refactor : LottoHandler package 위치 변경
